+++ keyboards/handwired/polykybd/base/text_helper.h
@@ -1,5 +1,8 @@
+// Copyright 2025 thpoll83
+// SPDX-License-Identifier: GPL-2.0-or-later
+
 #pragma once
 
 #include QMK_KEYBOARD_H
 
-const uint16_t* get_led_matrix_text(uint8_t rgb_mode);
+const uint32_t* get_led_matrix_text(uint8_t rgb_mode);
