+++ keyboards/handwired/polykybd/lang/lang_lut_ext.h
@@ -1,6 +1,9 @@
+// Copyright 2025 thpoll83
+// SPDX-License-Identifier: GPL-2.0-or-later
+
 #pragma once
 
-extern const uint16_t* latin_ex_map[26*2][10];
+extern const uint32_t* latin_ex_map[26*2][10];
 
 //extern const int8_t poly_settings [4][NUM_LANG * 4];
